+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Header component](components/Header.tsx)
 - [Custom hooks file](utils/hooks.ts) to utils directory
 - [Custom login page](pages/login.tsx)
+- [Default Avatar](components/DefaultAvatar.tsx) based on bootstrap icon 'person-fill'
 
 ### Changed